chore: 빌드 에러 수정 및 layout에서 권한 체크
