+++ app/(tabs)/_layout.tsx
@@ -1,5 +1,5 @@
-import { Tabs } from "expo-router";
-import React from "react";
+import { router, Tabs } from "expo-router";
+import React, { useEffect } from "react";
 import { Platform } from "react-native";
 
 import { HapticTab } from "@/components/HapticTab";
@@ -7,9 +7,16 @@ import { IconSymbol } from "@/components/ui/IconSymbol";
 import TabBarBackground from "@/components/ui/TabBarBackground";
 import { Colors } from "@/constants/Colors";
 import { useColorScheme } from "@/hooks/useColorScheme";
+import { useAuthStore } from "@/store/authStore";
 
 export default function TabLayout() {
   const colorScheme = useColorScheme();
+  const isLoggedIn = useAuthStore((state) => state.isLoggedIn);
+  useEffect(() => {
+    if (router && !isLoggedIn) {
+      router.replace("/login");
+    }
+  }, [isLoggedIn]);
 
   return (
     <Tabs
@@ -39,8 +46,10 @@ export default function TabLayout() {
       <Tabs.Screen
         name="myPage"
         options={{
-          title: 'MY',
-          tabBarIcon: ({ color }) => <IconSymbol size={28} name="paperplane.fill" color={color} />,
+          title: "MY",
+          tabBarIcon: ({ color }) => (
+            <IconSymbol size={28} name="paperplane.fill" color={color} />
+          ),
         }}
       />
     </Tabs>
